refactor names of classes and methods
