+++ .gitignore
@@ -1,2 +1,2 @@
-./.idea
-./target
+/.idea
+/target
