+++ src/main/java/edu/kis/vh/nursery/HanoiRhymer.java
@@ -1,6 +1,6 @@
 package edu.kis.vh.nursery;
 
-public class HanoiRhymer extends defaultCountingOutRhymer {
+public class HanoiRhymer extends DefaultCountingOutRhymer {
 
 int totalRejected = 0;
 
@@ -9,9 +9,9 @@ int totalRejected = 0;
 	}
 
 	public void countIn(int in) {
-	if (!callCheck() && in > peekaboo())
-			totalRejected++;
-			else
-				super.countIn(in);
+		if (!callCheck() && in > peekaboo())
+				totalRejected++;
+		else super.countIn(in);
 	}
+//	przełączanie między otwartymi kartami z plikami
 }
